Add work-in-progress day four part two code
I'm not sure why my code isn't giving the correct answer. It spits out
204, which is apparently too low. I've started debugging by printing
out all of the detected anagrams in an attempt to find any that aren't
actually anagrams, but it's already really late so I can't be bothered
to look through them all. This commit is just here so I don't have to
worry about moving the code over to my laptop in case I want to work
on it at school.