Add day four part two
* Add work-in-progress day four part two code

I'm not sure why my code isn't giving the correct answer. It spits out
204, which is apparently too low. I've started debugging by printing
out all of the detected anagrams in an attempt to find any that aren't
actually anagrams, but it's already really late so I can't be bothered
to look through them all. This commit is just here so I don't have to
worry about moving the code over to my laptop in case I want to work
on it at school.

* Check all letters of both words in isAnagram()

The isAnagram() function only looped through the letters in the first
word. This meant that in edge cases where the second word contained
all the letters of the first word and some others were detected as
anagrams.

My fix is to loop through all letters of both words. This means that
there is a redundant for loop for the vast majority of the words. I
originally thought to check if the tables were of the same size, but
Lua's table size operator ('#') only works on tables with integer
sequences for keys.

A possible optimization would be to store the lengths of the words,
but this code is just quick and dirty to get the answer. Now that I
think of it, I could probably just test the lengths of the words using
the words table I kept in order to print the words. Maybe another day.

* Remove debug printing