Add work-in-progress day six part one code
I started this at 11 and only just finished it. One of the people on
my private leaderboard finished within the first ten minutes, so it's
clearly not worth it to stay up until the new puzzle is released
because I can't compete anyway.

I'm really tired from not getting enough sleep the past few nights, so
I'm just going to put this here and clean it up tomorrow/another time.
The code works fine, just like I expect it to after I clean it up. But
there is a lot of debugging code that can be removed.